migrate to global Modal for Env Controls
